% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AraMetRoots}
\alias{AraMetRoots}
\title{A metabolite data set from Arabidopsis roots by GC-TOF/MS}
\description{
A metabolite data set. The Arabidopsis metabolome of the roots
of individual WT plants and the mto1 and tt4 mutants were
analyzed by GC-TOF/MS.
}
\details{
53 root samples (WT, n = 17; mto1 n = 16; and tt4, n = 20,
biological replicates).

A matrix containing  59 metabolites (rows) and 53 observations
(columns).

MetaboLights accession no.: MTBLS45

For comparisons with data from aerial parts (Kusano, Fukushima et al. 2007)
we selected 59 commonly-detected metabolites in both datasets using
MetMask http://metmask.sourceforge.net.
}
\author{
Atsushi Fukushima
}
\references{
Atsushi Fukushima et al. BMC Syst Biol 2011 5:1.
}
\keyword{datasets}

