% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffCorr.R
\name{compcorr}
\alias{compcorr}
\title{Compare two correlation coefficients}
\usage{
compcorr(n1, r1, n2, r2)
}
\arguments{
\item{n1}{sample size under condition 1}

\item{r1}{correlation coefficient under condition 1}

\item{n2}{sample size under condition 2}

\item{r2}{correlation coefficient under condition 1}
}
\value{
list of result (diff and p-value)
}
\description{
Compare two correlation coefficients using Fisher's Z-transformation
}
\examples{
compcorr(10, 0.1, 10, 0.9)
}
\references{
http://www.fon.hum.uva.nl/Service/Statistics/Two_Correlations.html
http://support.sas.com/ctx/samples/index.jsp?sid=494
http://support.sas.com/ctx/samples/index.jsp?sid=494
}
\author{
Atsushi Fukushima
}
