\name{get.min.max}
\alias{get.min.max}
\title{Get minimum and maximum}
\usage{
  get.min.max(d)
}
\arguments{
  \item{d}{data matrix or data frame}
}
\value{
  list object of minimum value or maximum value in a data
}
\description{
  Get minimum and maximum
}
\examples{
get.min.max(iris[,1:2])
}
\author{
  Atsushi Fukushima
}

