% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{golub.df}
\alias{golub.df}
\title{A compiled expression data from Golub et al. (1999)}
\description{
Golub data set from CoXpress paper by Michael Watson.
}
\details{
This is a data.frame of AML/ALL leukaemia microarray 
study, originally from Golub et al Science 286:531 (1999). 
It comprises gene expression measurements from 38 tumor 
mRNA samples, including 27 cases of acute lymphoblastic 
leukemia (ALL) and 11 cases of acute myeloid leukemia 
(AML). The HU6800 Affymetrix array, which contains 6,800 
probe sets, was used. The dataset was filtered to remove 
genes with negative values in any sample, resulting in a 
final set of 2,568 genes.
}
\references{
Michael Watson, BMC Bioinformatics 2006 7:509.
}
\keyword{datasets}
