% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffCorr.R
\name{cluster.molecule}
\alias{cluster.molecule}
\title{Hierarchical clustering of molecules}
\usage{
cluster.molecule(
  data,
  method = "pearson",
  linkage = "average",
  absolute = FALSE
)
}
\arguments{
\item{data}{matrix or data frame}

\item{method}{c("pearson", "spearman", "kendall", "euclidean",
"maximum", "manhattan", "canberra", "binary", or "minkowski")}

\item{linkage}{c("average", "ward", "single", "complete", "mcquitty",
"median", "centroid")}

\item{absolute}{if TRUE, then 1-|COR| else 1-COR, default is FALSE}
}
\value{
an object of class 'hclust'
}
\description{
Cluster molecules
}
\examples{
cluster.molecule(as.matrix(t(iris[,1:4])), "pearson", "average")
}
\author{
Atsushi Fukushima
}
