% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffCorr-package.R
\docType{package}
\name{DiffCorr-package}
\alias{DiffCorr-package}
\alias{_PACKAGE}
\title{Differential correlations in omics datasets}
\description{
Analyze and visualize differential correlations in biological networks
}
\author{
\strong{Maintainer}: Atsushi Fukushima \email{afukushima@gmail.com}

Authors:
\itemize{
  \item Kozo Nishida \email{kozo.nishida@gmail.com}
}

}
\keyword{internal}
