% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DiffCorr.R
\docType{package}
\name{DiffCorr}
\alias{DiffCorr}
\alias{DiffCorr-package}
\title{Differential correlations in omics datasets}
\description{
Analyze and visualize differential correlations in biological networks
}
\details{
\tabular{ll}{
Package: \tab DiffCorr \cr
Type: \tab Package \cr
Version: \tab 0.4.1 \cr
Date: \tab 2015-03-31 \cr
Depends: \tab igraph, pcaMethods, fdrtool \cr
License: \tab GPL (>=3) \cr
LazyLoad: \tab yes \cr
}

'
}
\author{
Atsushi Fukushima, Kozo Nishida
}

